Prevent tuple index out of range error
The original line:

```
if m.shape[1] > 3:
```

will error out with `IndexError: tuple index out of range`. It's an error I get running the example program example2_cdpsPlots.py, for example. If you preface it with the conditional `if len(m.shape) > 1 and`, this will prevent the `tuple index out of range` error as Python will move on to the else statement as soon as the first conditional in a chained conditional is false.

Tested on Mac and Linux with example2_cdpsPlots.py.